+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 2.1.12
+
+- Internal changes (No user-facing changes)
+  - Strongly type "collection".
+  - Add descriptions to return keys of `decode.details` function.
+
 ## 2.1.11
 
 - Add 571-581